Rude-Chat-flask-agents/Python-MCP-Server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.11.0
azure-ai-textanalytics==5.3.0
azure-search-documents==11.5.1
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-kusto-data==4.4.0
azure-kusto-ingest==4.4.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.0.0
semantic-kernel==1.33.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.67.0
eventlet==0.36.1
gunicorn==23.0.0

azure-storage-blob>=12.20.0
azure-ai-formrecognizer>=3.3.0
azure-core>=1.30.0

# Application Insights for telemetry and logging - specific compatible versions
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.2
opencensus-ext-logging==0.1.1

